Renamed workout repository for consistency. Removed redundant lines in gradle file.
