+++ data/src/main/kotlin/com/dvinc/workouttimer/data/repository/workout/mock/WorkoutDataRepositoryMock.kt
@@ -3,14 +3,14 @@
  * All rights reserved.
  */
 
-package com.dvinc.workouttimer.data.repository.workout
+package com.dvinc.workouttimer.data.repository.workout.mock
 
 import com.dvinc.workouttimer.domain.model.workout.Workout
 import com.dvinc.workouttimer.domain.repository.workout.WorkoutRepository
 import io.reactivex.Flowable
 import javax.inject.Inject
 
-class WorkoutDataRepository @Inject constructor() : WorkoutRepository {
+class WorkoutDataRepositoryMock @Inject constructor() : WorkoutRepository {
 
     override fun obtainWorkouts(): Flowable<List<Workout>> {
         return Flowable.fromArray(generateMocks())
