+++ domain/build.gradle
@@ -25,10 +25,6 @@ android {
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
     }
-
-    androidExtensions {
-        experimental = true
-    }
 }
 
 dependencies {
